AVRO-4023: Move `IdlUtils` into `org.apache.avro.idl`
Noticed this when building the distributions.

I think it makes sense to move it just into the `idl` namspace.

This package hasn't been released before, so we don't break any
existing APIs.

```
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-javadoc-plugin:3.8.0:jar (module-javadocs) on project avro-idl: MavenReportException: Error while generating Javadoc:
[ERROR] Exit code: 2
[ERROR] error: No source files for package org.apache.avro.util
[ERROR] 1 error
[ERROR] Command line was: /usr/lib/jvm/java-21-openjdk-arm64/bin/javadoc -J-Duser.language= -J-Duser.country= -Xdoclint:none @options @packages
[ERROR]
[ERROR] Refer to the generated Javadoc files in '/home/fokko.driesprong/avro/lang/java/idl/target/apidocs' dir.
[ERROR]
[ERROR] -> [Help 1]
[ERROR]
[ERROR] To see the full stack trace of the errors, re-run Maven with the -e switch.
[ERROR] Re-run Maven using the -X switch to enable full debug logging.
[ERROR]
[ERROR] For more information about the errors and possible solutions, please read the following articles:
[ERROR] [Help 1] http://cwiki.apache.org/confluence/display/MAVEN/MojoExecutionException
[ERROR]
[ERROR] After correcting the problems, you can resume the build with the command
[ERROR]   mvn <args> -rf :avro-idl
```
